feat: NetworkManager 정리
NetworkManager, ReadyNetworkManager 분리 -> 상속을 통해 중복 함수 제거
프리펩 변경 및 적용
불필요 씬 제거
변경점 각 씬 적용
