+++ Assets/Resources/Ready.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3ea6b803a47636346a207b5200871c18
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
