--- Assets/Scenes/ReadyScene_Spawn.unity.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-<<<<<<<< HEAD:Assets/Scenes/Multiplay.unity.meta
-guid: eafc1fa8bf11f0342bc3aa657cc39862
-========
-guid: 5efd180d84b275f47bea0b6036f06cfc
->>>>>>>> d20fb085feca69246b6f50e9132596a6ba5d30f6:Assets/Scenes/GambleScene.unity.meta
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
