+++ Assets/Scripts/Network/NetworkPlayer.cs
@@ -19,16 +19,12 @@ public class NetworkPlayer : MonoBehaviour
     private Transform leftHandRig;
     private Transform rightHandRig;
 
-    private GameObject localPlayer;
-
     void Start()
     {
         photonView = GetComponent<PhotonView>();
 
         if (photonView.IsMine)
         {
-            localPlayer = Instantiate(Resources.Load<GameObject>("Local Player"), new Vector3(1.8f, -2.8f, -1.2f), Quaternion.identity);
-            GambleManager.SetPlayer(localPlayer);
             foreach (var item in GetComponentsInChildren<Renderer>())
             {
                 item.enabled = false;
