+++ Assets/Scripts/Ready.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 722b345e75819a746b29fa3a60388496
+guid: 581d792b185080d49b23874ea948cb29
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
