+++ UserSettings/EditorUserSettings.asset
@@ -18,22 +18,22 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c72051a03113a20293f7e38271427fb
       flags: 0
     RecentlyUsedScenePath-4:
-      value: 22424703114646680e0b0227036c6d15171301192f2d2835633c133af6f9
-      flags: 0
-    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c78111b15142f1f2b233e28694c7df7ee3d2cfb
       flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c6d15171301192f2d28356d785326ece92021
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-6:
       value: 22424703114646680e0b0227036c6c0918142e2b3e212732212c0e7df7ee3d2cfb
+      flags: 0
+    RecentlyUsedScenePath-7:
+      value: 22424703114646680e0b0227036c6d15171301192f2d2835121a0d32f5ee7a2decee22f0
       flags: 0
     RecentlyUsedScenePath-8:
       value: 22424703114646680e0b0227036c78111b15142f1f2b233e2867083debf42d
       flags: 0
     RecentlyUsedScenePath-9:
-      value: 22424703114646680e0b0227036c6d15171301192f2d2835121a0d32f5ee7a2decee22f0
+      value: 22424703114646680e0b0227036c6d15171301192f2d2835633c133af6f9
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
